Fix MSVC C2668 ambiguous overload error for isnan/isinf/isfinite

+++ src/core_new/tensor/internal/tensor_functors.hpp
@@ -397,7 +397,7 @@ namespace lfs::core {
                 // Use unqualified name in device code - resolves to CUDA math function
                 return isnan(static_cast<float>(x)) ? 1 : 0;
 #else
-                return std::isnan(x) ? 1 : 0;
+                return std::isnan(static_cast<double>(x)) ? 1 : 0;
 #endif
             }
         };
@@ -409,7 +409,7 @@ namespace lfs::core {
                 // Use unqualified name in device code - resolves to CUDA math function
                 return isinf(static_cast<float>(x)) ? 1 : 0;
 #else
-                return std::isinf(x) ? 1 : 0;
+                return std::isinf(static_cast<double>(x)) ? 1 : 0;
 #endif
             }
         };
@@ -421,7 +421,7 @@ namespace lfs::core {
                 // Use unqualified name in device code - resolves to CUDA math function
                 return isfinite(static_cast<float>(x)) ? 1 : 0;
 #else
-                return std::isfinite(x) ? 1 : 0;
+                return std::isfinite(static_cast<double>(x)) ? 1 : 0;
 #endif
             }
         };
